App.vue: 기존 App.vue를 App_T.vue로 분리, 새 App.vue 추가
